requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
Django==5.0
django-daraja==1.3.0
django-mpesa==2.0.14
django-picklefield==3.2
djangorestframework==3.15.1
gunicorn==21.2.0
idna==3.7
jinxed==1.2.1
packaging==23.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.7.1
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
whitenoise==6.5.0
django-cors==0.1
django-cors-headers==4.3.1
django-currentuser==0.7.0
django-ext==0.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.3.0
PyJWT==2.8.0
requests==2.32.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
